Split tests into multiple modules.
